Modify index and styles
Edited the id's for the subsections of the content div to be better
styled. Added links for the roboto font in the head of index. Adjusted
body styling so that the page would only be a certain width and have
the imported roboto font. Added html styling in the css doc in order
to have the sides of the page be black. Added styling to first and
the button in first so that they would go together nicely. Renamed
the other subsection id's in the css file so they match the html file.
Changed background color of other subsections in content to see them
easier for later use.