Modify index, styles, and README
Added new jpg to the parent file and attributed the author in the
README file. Removed hard width from body to give a more responsive
design. Adjusted second sub-container in content in order to set it
up for later styling. Changed div to header in second sub-section of
content for proper emphasis. Added images to second content section
for later styling.