Modify index and styles body section 2
Got rid of unnecessary link to bold text in index. Added styling to
the second section in body to make it more like the design. Adjusted
padding in first section of body to make the page more like the design.
Adjusted styling of card and card img so that the cards would adjust
based on the amount of content within them.